Add drag handlers for NumberInput
