+++ packages/components/src/TextInput/__stories__/TextInputArrows.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { View } from '../../View';
+import { VStack } from '../../VStack';
+
+export const TextInputArrows = ({ decrement, increment }) => {
+	return (
+		<VStack spacing={0}>
+			<View as="button" onClick={increment}>
+				UP
+			</View>
+			<View as="button" onClick={decrement}>
+				DOWN
+			</View>
+		</VStack>
+	);
+};
