+++ tooling/vscode-snippets/package.json
